+++ source/MonoGame.Extended.Gui/Cursor.cs
@@ -1,5 +1,5 @@
 using Microsoft.Xna.Framework;
-using MonoGame.Extended.TextureAtlases;
+using MonoGame.Extended.Graphics;
 
 namespace MonoGame.Extended.Gui
 {
