+++ source/MonoGame.Extended.Gui/Element.cs
@@ -3,7 +3,7 @@ using System.ComponentModel;
 using System.Reflection;
 using System.Text.Json.Serialization;
 using Microsoft.Xna.Framework;
-using MonoGame.Extended.TextureAtlases;
+using MonoGame.Extended.Graphics;
 
 namespace MonoGame.Extended.Gui
 {
