+++ source/MonoGame.Extended/Graphics/NinePatchRegion2D.cs
@@ -1,7 +1,7 @@
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
-namespace MonoGame.Extended.TextureAtlases
+namespace MonoGame.Extended.Graphics
 {
     public class NinePatchRegion2D : TextureRegion2D
     {
