+++ source/MonoGame.Extended/Serialization/NinePatchRegion2DJsonConverter.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using MonoGame.Extended.TextureAtlases;
+using MonoGame.Extended.Graphics;
 
 namespace MonoGame.Extended.Serialization;
