+++ source/MonoGame.Extended/Sprites/SpriteSheetAnimation.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using MonoGame.Extended.Graphics;
 using MonoGame.Extended.TextureAtlases;
 
 namespace MonoGame.Extended.Sprites
