+++ source/MonoGame.Extended/TextureAtlases/TextureAtlasExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
+using MonoGame.Extended.Graphics;
 
 namespace MonoGame.Extended.TextureAtlases
 {
